requirements_enhanced.txt
# Jarvis AI - Enhanced Requirements
# Core dependencies for privacy-first modular AI development assistant

# Web Framework
streamlit>=1.30.0

# HTTP and API
requests>=2.31.0

# Configuration Management
pyyaml>=6.0

# Security and Authentication
cryptography>=40.0.0
bcrypt>=4.0.0

# System Monitoring
psutil>=5.9.0

# Data Processing
plotly>=5.14.0

# Search Integration
duckduckgo-search>=4.0.0

# Testing Framework
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Lang Family - AI Agent Framework
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.30
langsmith>=0.0.80

# Web Automation (Optional but recommended)
playwright>=1.35.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional but recommended for enhanced functionality
# Uncomment as needed:

# Git Integration
# gitpython>=3.1.0

# Data Science (for advanced analytics)
# pandas>=2.0.0
# numpy>=1.24.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Jupyter Integration
# notebook>=6.5.0
# ipywidgets>=8.0.0

# Database (if using SQLite alternatives)
# sqlalchemy>=2.0.0

# API Development
fastapi>=0.100.0
uvicorn>=0.22.0

# Advanced NLP (for enhanced text processing)
# spacy>=3.6.0
# nltk>=3.8.0

# Image Processing (for vision capabilities)
pillow>=10.0.0
# opencv-python>=4.8.0

# Machine Learning (for advanced features)
# scikit-learn>=1.3.0
# transformers>=4.30.0

# Cloud Integration (for external services)
# boto3>=1.28.0  # AWS
# azure-storage-blob>=12.17.0  # Azure
# google-cloud-storage>=2.10.0  # Google Cloud